[kpart.cpp] Remove hardcoded MEM_TRAFFIC_MAX
The maximum value of DRAM traffic (in bytes) is now retured by
CMTController.getMemTrafficMax(). As the comment there explains, this function
needs to hardcode the current width of IA32_QM_CTR.data since this cannot be
programatically determined. However, the multiplier is retrieved by querying
CPUID.(EAX=0FH, ECX=1H).EBX, instead of being hardcoded to 1<<15 (the value of
the multiplier for Xeon D-1540).